+++ frontend/admin/src/js/components/context/index.ts
@@ -1,0 +1,12 @@
+import AuthenticationProvider from "./AuthenticationProvider";
+import AuthorizationProvider from "./AuthorizationProvider";
+import ContextContainer from "./ContextContainer";
+import IntlContainer, { getMessages } from "./IntlContainer";
+
+export default ContextContainer;
+export {
+  AuthenticationProvider,
+  AuthorizationProvider,
+  IntlContainer,
+  getMessages,
+};
