+++ frontend/admin/src/js/dashboard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import { PoolDashboard } from "./components/PoolDashboard";
-import ContextContainer from "./components/ContextContainer";
+import ContextContainer from "./components/context/ContextContainer";
 
 ReactDOM.render(
   <ContextContainer>
